AGS: Fixed compilation error in ags.h (MVS-2013), some warning fixes
fixed compilation error in ags.h because NewInteractionEvent wasn't recognised as a class, gamefile.h has been included
fixed compilation error in pathfinder.cpp sqrt() in line 236 now takes double
fixed a warning in scripting.h DialogTopic is a struct not a class
fixed a warning in utils.cpp degree casted to float

+++ engines/ags/pathfinder.cpp
@@ -233,7 +233,7 @@ bool PathFinder::findNearestWalkableArea(const Graphics::Surface &tempMask, int 
 			if (*(const byte *)tempMask.getBasePtr(x, y) != 232)
 				continue;
 
-			uint distance = (uint)sqrt((x - _dest.x) * (x - _dest.x) + (y - _dest.y) * (y - _dest.y));
+			uint distance = (uint)sqrt( (double)(x - _dest.x) * (x - _dest.x) + (y - _dest.y) * (y - _dest.y));
 			if (distance < nearest) {
 				nearest = distance;
 				best = Common::Point(x, y);
